bgpd: Fix integer truncation of SRLG count when parsing SRLG TLV
`count` is declared as `uint8_t` but receives `length / 4` where
`length` is a `uint16_t`.  For `length >= 1024`, the division result
exceeds 255, which cannot be stored in `count` since `uint8_t` holds
at most 255, and the value silently truncates.  The result is then
stored in `attr->srlg_count`, also `uint8_t`, which would re-truncate
it.

Widen `count` in `parse_srlg()` and `srlg_count` in `struct
bgp_ls_attr` to `uint16_t`.

+++ bgpd/bgp_ls_nlri.c
@@ -3269,7 +3269,7 @@ static int parse_igp_metric(struct stream *s, uint16_t length, struct bgp_ls_att
  */
 static int parse_srlg(struct stream *s, uint16_t length, struct bgp_ls_attr *attr)
 {
-	uint8_t count;
+	uint16_t count;
 	int i;
 
 	if (length % 4 != 0) {
